+++ src/NavButtons/view.js
@@ -1,5 +1,8 @@
 import { NAV_BUTTON_ITEMS } from './constants.js';
-import { createElements, createContainer } from '../CreateElementUtils.js';
+import {
+  createElements,
+  createContainer,
+} from '../utils/createElementUtils.js';
 
 export default class NavButtons {
   constructor() {
