+++ src/PurchaseProducts/view.js
@@ -1,4 +1,4 @@
-import { createContainer } from '../CreateElementUtils.js';
+import { createContainer } from '../utils/createElementUtils.js';
 import {
   STRING_COIN_10,
   STRING_COIN_100,
